[CB-173] test, fix : 테스트 수정
################
[CB-274]
- 사용자 존재 확인과 모든 반려동물 불러오는 API 수정에 따른 테스트 수정
