+++ src/test/java/com/pinkdumbell/cocobob/domain/user/UserRepositoryTest.java
@@ -3,6 +3,8 @@ package com.pinkdumbell.cocobob.domain.user;
 import com.pinkdumbell.cocobob.config.JpaAuditingConfig;
 import com.pinkdumbell.cocobob.domain.pet.Pet;
 import com.pinkdumbell.cocobob.domain.pet.PetRepository;
+import com.pinkdumbell.cocobob.domain.pet.breed.Breed;
+import com.pinkdumbell.cocobob.domain.pet.breed.BreedRepository;
 import com.pinkdumbell.cocobob.domain.pet.image.PetImage;
 import com.pinkdumbell.cocobob.domain.pet.image.PetImageRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -15,7 +17,10 @@ import org.assertj.core.api.Assertions;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @DataJpaTest
 @Import(JpaAuditingConfig.class)
@@ -26,6 +31,8 @@ public class UserRepositoryTest {
     PetRepository petRepository;
     @Autowired
     PetImageRepository petImageRepository;
+    @Autowired
+    BreedRepository breedRepository;
     @PersistenceContext
     EntityManager em;
 
@@ -78,5 +85,36 @@ public class UserRepositoryTest {
         Assertions.assertThat(result.get().getEmail()).isEqualTo(email);
         Assertions.assertThat(result.get().getPets().size()).isEqualTo(2);
         Assertions.assertThat(result.get().getPets().get(0).getName()).isEqualTo(name);
+    }
+
+    @Test
+    @DisplayName("사용자 이메일을 통해 등록한 반려동물을 불러오는 메서드 테스트")
+    void testFindAllByUserIdWithBreed() {
+        String email = "test@test.com";
+        User user = userRepository.save(User.builder().email(email).build());
+        Breed breed = breedRepository.save(Breed.builder()
+                .name("진돗개")
+                .build());
+        petRepository.save(Pet.builder()
+                .user(user)
+                .breed(breed)
+                .name("코코")
+                .build());
+        petRepository.save(Pet.builder()
+                .user(user)
+                .breed(breed)
+                .name("키키")
+                .build());
+
+        em.flush();
+        em.clear();
+
+        Optional<User> result = userRepository.findUserByEmailWithPetDetail(user.getEmail());
+        List<Pet> pets = result.get().getPets();
+        Assertions.assertThat(pets.size()).isEqualTo(2);
+        Assertions.assertThat(pets.stream().map(Pet::getName).collect(Collectors.toList()))
+                .isEqualTo(Arrays.asList("코코", "키키"));
+        Assertions.assertThat(pets.stream().map(Pet::getBreed).map(Breed::getName).collect(Collectors.toList()))
+                .isEqualTo(Arrays.asList("진돗개", "진돗개"));
     }
 }
